requirements.txt
prometheus_client
pyyaml
requests
python-dateutil
python-json-logger
